# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Library Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "tlionnel@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get of Authors",
                "description": "Returns all authors",
                "operationId": "getAllAuthor",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Add an author",
                "description": "Add an author to the library",
                "operationId": "AddAuthor",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get One Author",
                "description": "Returns one author and associated authors and genre by his ID.",
                "operationId": "getOneAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by author id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete One Author",
                "description": "Delete one author by his ID.",
                "operationId": "DeleteAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete author by id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tests"
                ],
                "summary": "Update an author by his ID",
                "description": "Update an author to the library",
                "operationId": "UpdateAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get List Of Books",
                "description": "Returns all books and associated authors and genre.",
                "operationId": "getAllBooks",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tests"
                ],
                "summary": "Add a book",
                "description": "Add a book to the library",
                "operationId": "AddBook",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": true
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "author id",
                        "required": true
                    },
                    {
                        "name": "publication_year",
                        "in": "query",
                        "description": "publication year",
                        "required": true
                    },
                    {
                        "name": "pages_nb",
                        "in": "query",
                        "description": "number of pages",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Tests"
                ],
                "summary": "Get One Book",
                "description": "Returns one book and associated authors and genre by his ID.",
                "operationId": "getOneBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tests"
                ],
                "summary": "Delete One Book",
                "description": "Delete a book by is ID",
                "operationId": "deleteBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "delete book by id",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tests"
                ],
                "summary": "Update one book",
                "description": "Update a book by his ID",
                "operationId": "updateBook",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "search by book id",
                        "required": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Title",
                        "required": false
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Description",
                        "required": false
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "author id",
                        "required": false
                    },
                    {
                        "name": "publication_year",
                        "in": "query",
                        "description": "publication year",
                        "required": false
                    },
                    {
                        "name": "pages_nb",
                        "in": "query",
                        "description": "number of pages",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}